services/inference/requirements.txt
flask
redis
prometheus_client
pyyaml
joblib
numpy
